route: Fixing unregistration of fib entries
FaceMap was constructed from the face result of face registration.
But result gives a diffrent uri (instead of maia.cs.memphis.edu
it gives 141.225.11.107:6363). When unregistering it was not
working as NLSR tries to compare face uri of naming format. Fixed
that issue. Now unregistration of name prefixes works fine.

Change-Id: Iafb3790113f735b1bb0ea68360bf70d09ec54243